// src/models/infouser.ts
export interface InfoUser {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    confirm_password: string;
    type?: string;
}

export interface InfoUserValidation {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    confirm_password: string;
    type?: string;
}